Change core/store RowIter type from Iterator to Stream (#1419)
#### Background:
Currently, the return type of the data retrieval in the storage layer of the gluesql project is an `Iterator`. This design has limitations when integrating with asynchronous storage systems like Redis, where using a `Stream` would be more appropriate.

#### Problem:
The current `Iterator` return type doesn't allow us to pass data as a stream directly from storage layers that support streaming, such as Redis. As a result, the performance and efficiency of the system can be compromised.

#### Proposed Solution:
It would be beneficial to update the `RowIter` type from `Iterator` to `Stream` to enable better integration with storage systems that support asynchronous data retrieval and streaming.

#### Benefits:
- Improved performance by leveraging async capabilities of storage systems.
- More flexibility in storage implementations, allowing for more efficient data retrieval methods.
- Easier to integrate with modern async libraries and frameworks.